+++ src/layouts/BaseLayout.astro
@@ -7,7 +7,7 @@ import Footer from "~/components/Footer.astro";
 import Header from "~/components/Header.astro";
 import { AUTHOR, DIMENSIONS, SITE } from "~/config.ts";
 import { formatDateIso } from "~/date.ts";
-import { getFavicon } from "~/image.ts";
+import { getFavicon } from "~/pages/favicon.ico.ts";
 import { getUrl } from "~/url";
 
 interface Props {
