+++ src/pages/favicon.ico.ts
@@ -1,6 +1,8 @@
+import type { GetImageResult } from "astro";
 import type { Buffer } from "node:buffer";
 import sharp from "sharp";
 import ico from "sharp-ico";
+import { AVATAR, getImage } from "~/site.astro";
 
 async function getFaviconIco(): Promise<Buffer> {
   const buffer = await sharp("src/images/me.png")
@@ -8,6 +10,15 @@ async function getFaviconIco(): Promise<Buffer> {
     .toFormat("png")
     .toBuffer();
   return ico.encode([buffer]);
+}
+
+export async function getFavicon(size?: number): Promise<GetImageResult> {
+  return await getImage({
+    src: AVATAR,
+    width: size,
+    height: size,
+    format: "png",
+  });
 }
 
 export async function GET(): Promise<Response> {
