+++ src/scripts/build-photos.ts
@@ -26,44 +26,42 @@ interface PhotoData {
   license: string;
 }
 
-async function extractMetadata(photo: string): Promise<void> {
+async function extractMetadata(photo: string): Promise<PhotoData> {
   const cover = `${PHOTOS_DIR}/${photo}/wide.jpg`;
-  await exiftool.read<PhotoTags>(cover).then(async (tags) => {
-    const data: PhotoData = {
-      wide: `content/photos/${photo}/wide.jpg`,
-      square: `content/photos/${photo}/square.jpg`,
-      title: tags.Headline ?? "",
-      description: tags.ImageDescription ?? "",
-      keywords: Array.isArray(tags.Keywords)
-        ? tags.Keywords
-        : [tags.Keywords ?? ""],
-      date: (tags.DateTimeOriginal instanceof ExifDateTime
+  const tags = await exiftool.read<PhotoTags>(cover);
+  return {
+    wide: `content/photos/${photo}/wide.jpg`,
+    square: `content/photos/${photo}/square.jpg`,
+    title: tags.Headline ?? "",
+    description: tags.ImageDescription ?? "",
+    keywords: Array.isArray(tags.Keywords)
+      ? tags.Keywords
+      : [tags.Keywords ?? ""],
+    date:
+      (tags.DateTimeOriginal instanceof ExifDateTime
         ? tags.DateTimeOriginal.toDate().toISOString()
         : tags.DateTimeOriginal) ?? "",
-      location: tags.Location ?? "",
-      city: tags.City ?? "",
-      state: tags.State ?? "",
-      country: tags.Country ?? "",
-      camera: `${tags.Make ?? ""} ${tags.Model ?? ""}`.replace(/\s+/g, " "),
-      lens: tags.LensModel ?? tags.Lens ?? "",
-      editing: Array.isArray(tags.History)
-        ? (tags.History.filter((item) =>
-          item.Action === "produced"
-        )[0]
-          .SoftwareAgent ?? "")
-        : "",
-      license: tags.License ?? "",
-    };
-    const json = JSON.stringify(data, null, 2);
-    await Deno.writeTextFile(`${PHOTOS_DIR}/${photo}/index.json`, json);
-  });
+    location: tags.Location ?? "",
+    city: tags.City ?? "",
+    state: tags.State ?? "",
+    country: tags.Country ?? "",
+    camera: `${tags.Make ?? ""} ${tags.Model ?? ""}`.replace(/\s+/g, " "),
+    lens: tags.LensModel ?? tags.Lens ?? "",
+    editing: Array.isArray(tags.History)
+      ? (tags.History.filter((item) => item.Action === "produced")[0]
+        .SoftwareAgent ?? "")
+      : "",
+    license: tags.License ?? "",
+  };
 }
 
 try {
   for await (const cover of expandGlob(`${PHOTOS_DIR}/*/wide.jpg`)) {
     const photo = dirname(cover.path).split("/").pop();
     if (!photo) throw new Error("Could not get photo name");
-    await extractMetadata(photo);
+    const data = await extractMetadata(photo);
+    const json = JSON.stringify(data, null, 2);
+    await Deno.writeTextFile(`${PHOTOS_DIR}/${photo}/index.json`, json);
   }
 } finally {
   await exiftool.end();
